src/fca_categorization/requirements.txt
alembic==1.13.3
anyio==4.6.2.post1
asgiref==3.8.1
bcrypt==4.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
coverage==7.6.4
cryptography==43.0.3
ecdsa==0.19.0
fastapi==0.68.2
fca_categorization==1.0.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
Mako==1.3.6
MarkupSafe==3.0.2
packaging==24.1
passlib==1.7.4
pluggy==1.5.0
psutil==6.1.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==1.10.18
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.16
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.14.2
tabulate==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.15.0
